Moved query code and tests into lambda function
