+++ lambdas/scheduler/tests/test_scheduler.py
@@ -2,11 +2,11 @@ import import_scheduler
 
 import schedule
 import testing_utils as utils
-from db_cred_setup import skip_if_creds_not_availbable
+from db_cred_setup import skip_if_creds_not_available
 from environment import environment
 
 
-@skip_if_creds_not_availbable
+@skip_if_creds_not_available
 def test_scheduler():
     granules = utils.load_testing_granules()
 
@@ -23,3 +23,22 @@ def test_scheduler():
 
     if 'local' in environment.maturity:
         utils.cache_results(email_packages)
+
+
+@skip_if_creds_not_available
+def test_get_users_by_ids(db):
+    ids = [37, 34]
+    users = db.get_users_by_ids(ids)
+
+    assert users
+    for user in users:
+        assert user.id in ids
+
+
+@skip_if_creds_not_available
+def test_get_enabled_subs(db):
+    enabled_subs = db.get_enabled_subs()
+
+    assert enabled_subs
+    for sub in enabled_subs:
+        assert sub.enabled is True
